calculator/requirements.txt
Pillow==9.2.0

sympy~=1.11.1